refactor: move Contract and Instruction to separate module (#520)
